feat(normalize): allow normalize to work with arrays
Options with 'normalize' and 'array' no longer fail when passing more than 1
path.
The previous behaviour of getter and setter is only applied when the option
doesn't use the 'array' flag.
In the future, we can use ES2015's proxies to also apply the same effect of the
getter and setter to all the elements in the array.
Note: For this to work, you need to set your option to 'array', it doesn't work
with implicit arrays.

Closes yargs/#430