feat(configuration): Allow to pass configuration objects to yargs-parser
This was brought to by the issue #435 in yargs, about aliases not being applied
correctly to values set by pkgConf.
For aliases to be handled correctly, even considering transitive relationships,
pkgConf values should be handled in yargs-parser, where most of the logic of
aliases is. So, the solution here is to pass the object from the key in
package.json to yargs-parser.
This is the first part of a solution that would later need yargs to explicitly
pass the object yargs-parser.
Here are some designs considerations I made:
    - I thought the best is to add a generic key to opts in yargs-parser, called
      'configObjects' that takes arbitrary objects and adds them to argv.
      That way it makes sense to people who use yargs-parser by itself.
    - Also, I thought it's fine for them to have almost the same priority that
      values set by a config file, after all, they are another way of
      configuration objects. An issue with that is that values from pkgConf
      were previously added as 'defaults', so they could have been
      overridden by env values before, but not anymore.