Do not parse numbers that evaluate to Infinity
Some numbers, like `87e3202`, are recognised as numbers by the
isNumber utility, which results on `yargs-parser` parsing them by default
and thus passing a meaningless `Infinity` value to the user.

Since parsing scientific notation numbers are supported on `yargs-parser` as a
feature, the approach of this patch is to only return true from
`isNumber` if the string indeed represents a number, but it doesn't
evaluate to `Infinity` when being parsed as such, therefore not breaking
any existing functionality on the module.

The use case where this bug showed up was passing (shorter) UUIDs as
command line arguments.

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>